+++ shared/core/data/src/commonMain/kotlin/ru/aleshin/studyassistant/core/data/repositories/SubjectsRepositoryImpl.kt
@@ -168,21 +168,25 @@ class SubjectsRepositoryImpl(
         }
     }
 
-    override suspend fun transferData(direction: DataTransferDirection) {
+    override suspend fun transferData(direction: DataTransferDirection, mergeData: Boolean) {
         val currentUser = userSessionProvider.getCurrentUserId()
         when (direction) {
             DataTransferDirection.REMOTE_TO_LOCAL -> {
                 val allSubjectsFlow = remoteDataSource.fetchAllItems(currentUser)
                 val subjects = allSubjectsFlow.first().map { it.convertToLocal() }
 
-                localDataSource.offline().deleteAllItems()
+                if (!mergeData) {
+                    localDataSource.offline().deleteAllItems()
+                }
                 localDataSource.offline().addOrUpdateItems(subjects)
             }
             DataTransferDirection.LOCAL_TO_REMOTE -> {
                 val allSubjects = localDataSource.offline().fetchAllSubjects().first()
                 val subjectsRemote = allSubjects.map { it.convertToRemote(currentUser) }
 
-                remoteDataSource.deleteAllItems(currentUser)
+                if (!mergeData) {
+                    remoteDataSource.deleteAllItems(currentUser)
+                }
                 remoteDataSource.addOrUpdateItems(subjectsRemote)
 
                 localDataSource.sync().deleteAllItems()
